Split public space search into its own feature with separate files
